+++ src/lib/utilsServer/courier/getClient.ts
@@ -1,0 +1,7 @@
+import { CourierClient } from '@trycourier/courier'
+import { assert } from '@lib'
+
+export const getClient = () => {
+  const authorizationToken = assert(process.env.COURIER_API_KEY, 'COURIER_API_KEY')
+  return CourierClient({ authorizationToken })
+}
